Add developer tooling and fix static analysis defect findings
- Add justfile providing recipes for build, test, cppcheck, and clang-tidy
- Add .clang-tidy configured for C++11 and project include paths
- Add AGENTS.md guidelines for AI coding assistants and contributor instructions in BUILD.md
- Fix array bounds out-of-range indexing in symmetries_0.cpp, symmetries_1.cpp, symmetries_2.cpp
- Add [[noreturn]] attribute to bt_te() to aid static analyzer path pruning
- Fix potential null pointer dereferences and unsafe realloc patterns in Fl_Table.h and stlexportsolution.cpp
- Fix negative bit-shift and depth check logic in xml.cpp
- Fix signed/unsigned loop index in assembly.cpp and explicit string comparisons in assembler_0.cpp
- Add self-assignment check to vector3::operator= and unhide Fl_Double_Window::hide in multilinewindow.h

+++ src/lib/voxel_4.cpp
@@ -482,9 +482,9 @@ void voxel_4_c::getConnectionFace(int x, int y, int z, int n, double bevel, doub
 
 
 void voxel_4_c::calculateSize(float * x, float * y, float * z) const {
-  *x = (getX()+2)/3;
-  *y = (getY()+2)/3;
-  *z = (getZ()+2)/3;
+  *x = (float)((getX() + 2) / 3);
+  *y = (float)((getY() + 2) / 3);
+  *z = (float)((getZ() + 2) / 3);
 }
 
 void voxel_4_c::recalcSpaceCoordinates(float * x, float * y, float * z) const {
